refactor: add factory operator
